Wed, Aug 26, 2026, 1:33 PM +03:00

+++ app.c
@@ -45,7 +45,7 @@ int main(void)
 
                       {
 
-                              printf("%i\n", x/y);
+                              printf("%f\n",(float) x/y);
 }
